sensors: unloadProgram to load.go
move unloadProgram function to load.go
