+++ pkg/sensors/load.go
@@ -364,6 +364,25 @@ func createDir(bpfDir, mapDir string) {
 	os.Mkdir(mapDir, os.ModeDir)
 }
 
+func unloadProgram(prog *program.Program) {
+	log := logger.GetLogger().WithField("label", prog.Label).WithField("pin", prog.PinPath)
+
+	if !prog.LoadState.IsLoaded() {
+		log.Debugf("Refusing to remove %s, program not loaded", prog.Label)
+		return
+	}
+	if count := prog.LoadState.RefDec(); count > 0 {
+		log.Debugf("Program reference count %d, not unloading yet", count)
+		return
+	}
+
+	if err := prog.Unload(); err != nil {
+		logger.GetLogger().WithField("name", prog.Name).WithError(err).Warn("Failed to unload program")
+	}
+
+	log.Info("BPF prog was unloaded")
+}
+
 func UnloadAll(bpfDir string) {
 	for _, l := range AllPrograms {
 		unloadProgram(l)
